requirements.txt
pandas
openpyxl
numpy
requests
unidecode
selenium
beautifulsoup4
geopy
python-dotenv